+++ sentry_sdk/tracing_utils_python2.py
@@ -3,11 +3,13 @@ from functools import wraps
 import sentry_sdk
 from sentry_sdk._types import TYPE_CHECKING
 from sentry_sdk.consts import OP
-from sentry_sdk.tracing import Transaction
 from sentry_sdk.utils import qualname_from_function
+
 
 if TYPE_CHECKING:
     from typing import Any, Optional
+
+    from sentry_sdk.tracing import Transaction
 
 
 def start_child_span_decorator(func):
@@ -16,7 +18,7 @@ def start_child_span_decorator(func):
     Decorator to add child spans for functions.
 
     This is the Python 2 compatible version of the decorator.
-    Duplicated code from ``sentry_sdk.tracing_utils.start_child_span_decorator``.
+    Duplicated code from ``sentry_sdk.tracing_utils_python3.start_child_span_decorator``.
 
     See also ``sentry_sdk.tracing.trace()``.
     """
